% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{add_urls_to_scenarios}
\alias{add_urls_to_scenarios}
\title{Add urls to scenarios}
\usage{
add_urls_to_scenarios(scenarios, app_path, port = default_port)
}
\arguments{
\item{scenarios}{A list with scenarios (read from json)}

\item{app_path}{Relative path or an url for Shiny App.}

\item{port}{Port in which local Shiny App should be run.}
}
\description{
Add urls to scenarios
}
