% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{define_scenario}
\alias{define_scenario}
\title{Define scenario basing on template}
\usage{
define_scenario(label, config_dir = getOption("config_dir", default =
  default_config_dir), onBeforeScript = NULL, cookiePath = NULL,
  referenceUrl = NULL, readyEvent = NULL, readySelector = NULL,
  delay = 1000, hideSelectors = NULL, removeSelectors = list(),
  onReadyScript = NULL, keyPressSelectors = NULL,
  hoverSelector = NULL, hoverSelectors = NULL, clickSelector = NULL,
  clickSelectors = NULL, postInteractionWait = NULL,
  scrollToSelector = NULL, selectors = list("document"),
  selectorExpansion = NULL, misMatchThreshold = 0.2,
  requireSameDimensions = NULL)
}
\arguments{
\item{label}{Scenario id.}

\item{config_dir}{Relative path to structure config file (config.yaml) directory.}

\item{url}{URL to application on which tests should be performed.}

\item{onBeforeScriptcookiePath, referenceUrl, readyEvent, readySelector, delay, hideSelectors, removeSelectors, onReadyScript, keyPressSelectors, hoverSelector, hoverSelectors, clickSelector, clickSelectors, postInteractionWait, scrollToSelector, selectors, selectorExpansion, misMatchThreshold, requireSameDimensions}{Other parameters defining scenario. Please check https://github.com/garris/BackstopJS for details.}
}
\description{
Define scenario basing on template
}
