% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{edit_scenario}
\alias{edit_scenario}
\title{Edit scenario}
\usage{
edit_scenario(label, config_dir = getOption("config_dir", default =
  default_config_dir))
}
\arguments{
\item{label}{Scenario id.}

\item{config_dir}{Relative path to structure config file (config.yaml) directory.}
}
\description{
Edit scenario
}
