% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure.R
\name{make_structure}
\alias{make_structure}
\title{Create structure for tests}
\usage{
make_structure(dir = "tests/end2end", id = "app", env_script = FALSE,
  settings = FALSE)
}
\arguments{
\item{dir}{Directory in which structure should be created}

\item{id}{Subdirectory of dir. dir/id is the final directory for tests structure.}

\item{env_script}{If TRUE, then empty R script for setting local Shiny App environment is created. Not used yet.}

\item{settings}{If TRUE, then settings template for browser and test environment is created. Not used yet.}
}
\description{
Create structure for tests
}
