% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{open_report}
\alias{open_report}
\title{Open test report}
\usage{
open_report(config_dir)
}
\arguments{
\item{config_dir}{Relative path to structure config file (config.yaml) directory.}
}
\description{
Open test report
}
