% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{path_is_local}
\alias{path_is_local}
\title{Check if string is a local path or url}
\usage{
path_is_local(path)
}
\arguments{
\item{path}{Path or url}
}
\description{
Check if string is a local path or url
}
